Initial import
A rewrite of Hans Hübner's ACL-ZMQ, which was based on Vitaly
Mayatskikh CL-ZMQ.

This version uses a C thread to perform potentially blocking
operations, so the whole (green-threaded) Allegro Common Lisp image
does not get blocked.

The only thing left over from the original implementation is the
defconstants, so I'm going ahead and releasing this under my own
license.